requirements.txt
Flask
pymongo
werkzeug
tesseract
pillow
pytesseract
pdf2image